refactor rank
